Move algorithm functions to the BLs
